feat(silent): Cancel one instance of model change pipeline.
Added the ability for a protocol to set the silent property of a delta to true in order to prevent the model change pipeline from firing immediately after the change. This will not stop the model update from propagating throughout the application; it will just prevent it from being sent on a round trip back to the protocol.